requirements.in.txt
apscheduler
chromadb
ctransformers
epub2txt
fastapi
httpx
InstructorEmbedding
langchain
langchain-community
llama_cpp_python
msg-parser
nc_py_api
odfdo
odfpy
openpyxl
pandas
pypdf
docx2txt
protobuf
python-dotenv
python-multipart
python-pptx
ruamel.yaml
sentence-transformers
setuptools
sse_starlette
starlette_context
striprtf
torch
torchvision
transformers
unstructured @ git+https://github.com/kyteinsky/unstructured@d3a404cfb541dae8e16956f096bac99fc05c985b
unstructured-client
weaviate-client
xlrd
